+++ armada_backend/api_register.py
@@ -1,7 +1,10 @@
+import falcon
+
 from armada_backend import api_base, docker_client
+from armada_backend.exceptions import BadRequestException
+from armada_command.armada_utils import print_err
 from armada_command.consul.consul import consul_post, consul_query
 from armada_command.consul.kv import kv_set, kv_remove
-from armada_command.scripts.compat import json
 
 
 def _exists_service(service_id):
@@ -38,17 +41,21 @@ def register_service_in_consul(microservice_data):
         kv_remove(key)
 
 
-def local_port_to_external_port(microservice_id, microservice_local_port):
+def local_port_to_external_port(container_id, microservice_local_port):
     docker_api = docker_client.api()
-    docker_inspect = docker_api.inspect_container(microservice_id.split(':')[0])
-    external_port = int(docker_inspect['NetworkSettings']['Ports'][microservice_local_port][0]['HostPort'])
+    docker_inspect = docker_api.inspect_container(container_id)
+    container_ports = docker_inspect['NetworkSettings']['Ports']
+    if microservice_local_port not in container_ports:
+        raise BadRequestException(
+            'Container with id="{}" does not expose port "{}"'.format(container_id, microservice_local_port))
+    external_port = int(container_ports[microservice_local_port][0]['HostPort'])
     return external_port
 
 
 class Register(api_base.ApiCommand):
     def on_post(self, req, resp):
         try:
-            microservice_data = json.load(req.stream)
+            microservice_data = req.json
             register_service_in_consul(microservice_data)
             result = {'microservice_data': microservice_data}
         except Exception as e:
@@ -57,16 +64,17 @@ class Register(api_base.ApiCommand):
 
 
 class RegisterV1(api_base.ApiCommand):
-    def on_post(self, req, resp):
+    def on_post(self, req, resp, microservice_id):
         try:
-            input_json = json.load(req.stream)
+            input_json = req.json
+            print_err('input_json: {}'.format(input_json))
+            container_id = microservice_id.split(':')[0]
             microservice_data = {
-                'microservice_id': input_json['microservice_id'],
+                'microservice_id': microservice_id,
                 'microservice_name': input_json['microservice_name'],
                 'single_active_instance': bool(input_json.get('single_active_instance')),
                 'container_created_timestamp': input_json['container_created_timestamp'],
-                'microservice_port': local_port_to_external_port(input_json['microservice_id'],
-                                                                 input_json['microservice_local_port']),
+                'microservice_port': local_port_to_external_port(container_id, input_json['microservice_local_port']),
             }
             microservice_tags = []
             if input_json.get('microservice_env'):
@@ -76,7 +84,8 @@ class RegisterV1(api_base.ApiCommand):
             if microservice_tags:
                 microservice_data['microservice_tags'] = microservice_tags
             register_service_in_consul(microservice_data)
-            result = {'microservice_data': microservice_data}
+            resp.json = microservice_data
         except Exception as e:
-            return self.status_exception(resp, 'Could not register service.', e)
-        return self.status_ok(resp, {'result': result})
+            print_err(str(e))
+            resp.json = {'error': 'Could not register service: {}'.format(repr(e))}
+            resp.status = falcon.HTTP_400