+++ armada_backend/api_start.py
@@ -2,8 +2,8 @@ import six
 
 from armada_backend import api_base, docker_client
 from armada_backend.models.services import save_container
-from armada_backend.utils import shorten_container_id
 from armada_backend.models.ships import get_ship_name
+from armada_backend.utils import shorten_container_id
 from armada_command.consul.consul import consul_query