+++ armada_backend/api_version.py
@@ -5,4 +5,5 @@ from armada_backend import api_base
 
 class GetVersion(api_base.ApiCommand):
     def on_get(self, req, resp):
+        resp.content_type = 'text/plain'
         resp.body = os.environ.get("ARMADA_VERSION", "none")