+++ armada_backend/runtime_settings.py
@@ -1,7 +1,7 @@
 from armada_backend import consul_config
+from armada_backend.models.ships import get_ship_name, get_other_ship_ips
 from armada_backend.utils import get_current_datacenter, is_ship_commander, \
     setup_sentry
-from armada_backend.models.ships import get_ship_name, get_other_ship_ips
 from armada_command.dockyard import alias
 from armada_command.scripts.compat import json