+++ docker-containers/microservice/microservice/defines.py
@@ -1,0 +1,1 @@
+ARMADA_API_URL = 'http://172.17.0.1:8900'